Add workshop admin tests
