+++ rcjaRegistration/common/baseTests/__init__.py
@@ -1,3 +1,3 @@
-from .populateDatabase import createStates, createUsers, createSchools, createEvents, createTeams, createInvoices, createQuestionsAndResponses, createAssociationMemberships
+from .populateDatabase import createStates, createUsers, createSchools, createEvents, createTeams, createWorkshopAttendees, createInvoices, createQuestionsAndResponses, createAssociationMemberships
 from .adminPermissions import Base_Test_NotStaff, Base_Test_SuperUser, Base_Test_FullCoordinator, Base_Test_ViewCoordinator, GET_SUCCESS, GET_DENIED, GET_DENIED_ALL, ADDDELETE_PAGE_DENIED_VIEWONLY, POST_SUCCESS, POST_VALIDATION_FAILURE, POST_DENIED
 from .adminPermissions import Base as Base_Admin_Test
