+++ rcjaRegistration/common/baseTests/populateDatabase.py
@@ -7,6 +7,7 @@ from coordination.models import Coordinator
 from teams.models import Team, Student, HardwarePlatform, SoftwarePlatform
 from userquestions.models import Question, QuestionResponse
 from association.models import AssociationMember
+from workshops.models import WorkshopAttendee
 
 import datetime
 
@@ -174,6 +175,21 @@ def createEvents(self):
         venue=self.venue3_state2,
     )
 
+    self.state2_openWorkshop = Event.objects.create(
+        year = self.year,
+        state = self.state2,
+        name = 'State 2 Open Workshop',
+        eventType = 'workshop',
+        status = 'published',
+        competition_defaultEntryFee = 50,
+        startDate = (datetime.datetime.now() + datetime.timedelta(days=3)).date(),
+        endDate = (datetime.datetime.now() + datetime.timedelta(days=3)).date(),
+        registrationsOpenDate = (datetime.datetime.now() + datetime.timedelta(days=-10)).date(),
+        registrationsCloseDate = (datetime.datetime.now() + datetime.timedelta(days=1)).date(),
+        directEnquiriesTo = self.user_state1_super1,
+        venue=self.venue3_state2,
+    )
+
     # Divisions
 
     self.division1_state1 = Division.objects.create(name='Division 1', state=self.state1)
@@ -184,7 +200,7 @@ def createEvents(self):
 
     # Available Divisions
 
-    for event in ['state1_openCompetition', 'state1_openWorkshop', 'state1_closedCompetition1', 'state1_closedCompetition2', 'state1_pastCompetition', 'state2_openCompetition']:
+    for event in ['state1_openCompetition', 'state1_openWorkshop', 'state1_closedCompetition1', 'state1_closedCompetition2', 'state1_pastCompetition', 'state2_openCompetition', 'state2_openWorkshop']:
         setattr(self, f'availableDivision3_{event}', AvailableDivision.objects.create(event=getattr(self, event), division=self.division3))
         setattr(self, f'availableDivision4_{event}', AvailableDivision.objects.create(event=getattr(self, event), division=self.division4))
 
@@ -220,6 +236,41 @@ def createTeams(self):
         softwarePlatform=self.softwarePlatform,
     )
 
+def createWorkshopAttendees(self):
+    self.state1_event1_workshopAttendee1 = WorkshopAttendee.objects.create(
+        event=self.state1_openWorkshop,
+        mentorUser=self.user_state1_school1_mentor1,
+        school=self.school1_state1,
+        division=self.division3,
+        firstName='First 1',
+        lastName='Last 1',
+        yearLevel=10,
+        attendeeType='student',
+        gender='male,'
+    )
+    self.state1_event1_workshopAttendee2 = WorkshopAttendee.objects.create(
+        event=self.state1_openWorkshop,
+        mentorUser=self.user_state1_school1_mentor1,
+        school=self.school1_state1,
+        division=self.division3,
+        firstName='First 2',
+        lastName='Last 2',
+        yearLevel=11,
+        attendeeType='student',
+        gender='male',
+    )
+    self.state2_event1_workshopAttendee3 = WorkshopAttendee.objects.create(
+        event=self.state2_openWorkshop,
+        mentorUser=self.user_state1_school1_mentor1,
+        school=self.school1_state1,
+        division=self.division3,
+        firstName='First 3',
+        lastName='Last 3',
+        yearLevel=11,
+        attendeeType='student',
+        gender='male',
+    )
+
 def createInvoices(self):
     self.state1_event1_invoice1 = Invoice.objects.create(
         invoiceToUser=self.user_state1_school1_mentor1,
@@ -230,6 +281,17 @@ def createInvoices(self):
         invoiceToUser=self.user_state1_school1_mentor1,
         school=self.school1_state1,
         event=self.state2_openCompetition
+    )
+
+    self.state1_workshopEvent1_invoice1 = Invoice.objects.create(
+        invoiceToUser=self.user_state1_school1_mentor1,
+        school=self.school1_state1,
+        event=self.state1_openWorkshop
+    )
+    self.state2_workshopEvent1_invoice2 = Invoice.objects.create(
+        invoiceToUser=self.user_state1_school1_mentor1,
+        school=self.school1_state1,
+        event=self.state2_openWorkshop
     )
 
 def createQuestionsAndResponses(self):
